// src/types/store-json.ts
 export type SdkRequestType = AuthCustomerAuthProvider | AuthCustomerAuthProviderCallback | AuthCustomerAuthProviderRegister | AuthCustomerAuthProviderResetPassword | AuthCustomerAuthProviderUpdate | AuthSession | AuthTokenRefresh | StoreCarts | StoreCartsId | StoreCartsIdComplete | StoreCartsIdCustomer | StoreCartsIdLineItems | StoreCartsIdLineItemsLineId | StoreCartsIdPromotions | StoreCartsIdShippingMethods | StoreCartsIdTaxes | StoreCollections | StoreCollectionsId | StoreCurrencies | StoreCurrenciesCode | StoreCustomers | StoreCustomersMe | StoreCustomersMeAddresses | StoreCustomersMeAddressesAddressId | StoreOrders | StoreOrdersId | StoreOrdersIdTransferAccept | StoreOrdersIdTransferCancel | StoreOrdersIdTransferDecline | StoreOrdersIdTransferRequest | StorePaymentCollections | StorePaymentCollectionsIdPaymentSessions | StorePaymentProviders | StoreProductCategories | StoreProductCategoriesId | StoreProductTags | StoreProductTagsId | StoreProductTypes | StoreProductTypesId | StoreProducts | StoreProductsId | StoreRegions | StoreRegionsId | StoreReturnReasons
// ...
  export interface AuthCustomerAuthProvider {
    post: Post
  }
  
  export interface Post {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses
  }
  
  
  export interface Parameter {
    name: string
    in: string
    description: string
    required: boolean
    schema: Schema
  }
  
  export interface Schema {
    type: string
    example: string
  }
// ...
  export interface AuthCustomerAuthProviderCallback {
    post: Post2
  }
  
  export interface Post2 {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses2
  }
// ...
  export interface AuthCustomerAuthProviderRegister {
    post: Post3
  }
  
  export interface Post3 {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody2
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses3
  }
// ...
  export interface AuthCustomerAuthProviderResetPassword {
    post: Post4
  }
  
  export interface Post4 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody3
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses4
    "x-workflow": string
  }
// ...
  export interface AuthCustomerAuthProviderUpdate {
    post: Post5
  }
  
  export interface Post5 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody4
    "x-codeSamples": CodeSample[]
    security: Security[]
    tags: string[]
    responses: Responses5
  }
// ...
  export interface AuthSession {
    post: Post6
    delete: Delete
  }
  
  export interface Post6 {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    parameters: Parameter[]
    "x-authenticated": boolean
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses6
  }
// ...
  export interface Delete {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses7
  }
// ...
  export interface AuthTokenRefresh {
    post: Post7
  }
  
  export interface Post7 {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    parameters?: Parameter[]
    "x-authenticated": boolean
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses8
  }
// ...
  export interface StoreCarts {
    post: Post8
  }
  
  export interface Post8 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody5
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses9
    "x-workflow": string
  }
// ...
  export interface StoreCartsId {
    get: Get
    post: Post9
  }
  
  export interface Get {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses10
  }
// ...
  export interface Post9 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody6
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses11
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdComplete {
    post: Post10
  }
  
  export interface Post10 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    externalDocs: ExternalDocs
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses12
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdCustomer {
    post: Post11
  }
  
  export interface Post11 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses13
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdLineItems {
    post: Post12
  }
  
  export interface Post12 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody7
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses14
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdLineItemsLineId {
    post: Post13
    delete: Delete2
  }
  
  export interface Post13 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody8
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses15
    "x-workflow": string
  }
// ...
  export interface Delete2 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses16
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdPromotions {
    post: Post14
    delete: Delete3
  }
  
  export interface Post14 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody9
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses17
    "x-workflow": string
  }
// ...
  export interface Delete3 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses18
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdShippingMethods {
    post: Post15
  }
  
  export interface Post15 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody10
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses19
    "x-workflow": string
  }
// ...
  export interface StoreCartsIdTaxes {
    post: Post16
  }
  
  export interface Post16 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses20
    "x-workflow": string
  }
// ...
  export interface StoreCollections {
    get: Get2
  }
  
  export interface Get2 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    externalDocs: ExternalDocs
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses21
  }
// ...
  export interface StoreCollectionsId {
    get: Get3
  }
  
  export interface Get3 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    externalDocs: ExternalDocs
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses22
  }
// ...
  export interface StoreCurrencies {
    get: Get4
  }
  
  export interface Get4 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses23
  }
// ...
  export interface StoreCurrenciesCode {
    get: Get5
  }
  
  export interface Get5 {
    operationId: string
    summary: string
    description: string
    "x-authenticated": boolean
    parameters: Parameter[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses24
  }
// ...
  export interface StoreCustomers {
    post: Post17
  }
  
  export interface Post17 {
    operationId: string
    summary: string
    description: string
    externalDocs: ExternalDocs
    "x-authenticated": boolean
    parameters: Parameter[]
    requestBody: RequestBody11
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses25
    "x-workflow": string
    security: Security2[]
  }
// ...
  export interface StoreCustomersMe {
    get: Get6
    post: Post18
  }
  
  export interface Get6 {
    operationId: string
    summary: string
    "x-sidebar-summary": string
    description: string
    "x-authenticated": boolean
    externalDocs: ExternalDocs
    parameters: Parameter[]
    security: Security3[]
    "x-codeSamples": CodeSample[]
    tags: string[]
    responses: Responses26
  }
// ...
  export interface StoreCustomersMeAddresses {
    get: Get7
    post: Post19
  }
// ...
  export interface StoreCustomersMeAddressesAddressId {
    get: Get8
    post: Post20
    delete: Delete4
  }
// ...
  export interface StoreOrders {
    get: Get9
  }
// ...
  export interface StoreOrdersId {
    get: Get10
  }
// ...
  export interface StoreOrdersIdTransferAccept {
    post: Post21
  }
// ...
  export interface StoreOrdersIdTransferCancel {
    post: Post22
  }
// ...
  export interface StoreOrdersIdTransferDecline {
    post: Post23
  }
// ...
  export interface StoreOrdersIdTransferRequest {
    post: Post24
  }
// ...
  export interface StorePaymentCollections {
    post: Post25
  }
// ...
  export interface StorePaymentCollectionsIdPaymentSessions {
    post: Post26
  }
// ...
  export interface StorePaymentProviders {
    get: Get11
  }
// ...
  export interface StoreProductCategories {
    get: Get12
  }
// ...
  export interface StoreProductCategoriesId {
    get: Get13
  }
// ...
  export interface StoreProductTags {
    get: Get14
  }
// ...
  export interface StoreProductTagsId {
    get: Get15
  }
// ...
  export interface StoreProductTypes {
    get: Get16
  }
// ...
  export interface StoreProductTypesId {
    get: Get17
  }
// ...
  export interface StoreProducts {
    get: Get18
  }
// ...
  export interface StoreProductsId {
    get: Get19
  }
// ...
  export interface StoreRegions {
    get: Get20
  }
// ...
  export interface StoreRegionsId {
    get: Get21
  }
// ...
  export interface StoreReturnReasons {
    get: Get22
  }